Feature/WrenAPIimprovements (#162)
* [Add] improvements to wren math API

* [Fix] renamed and fixed math functions in player movement.wren

* [Changed] reorganized entity binding to another file

* [Add] shutdown function for main scripts

* [Add] wren binding to switch between main scripts

* [Fix] formatting and reverted back to using game as starting script
