+++ engine/main.cpp
@@ -46,6 +46,7 @@ int main(MAYBE_UNUSED int argc, MAYBE_UNUSED char* argv[])
 
     BindEngineAPI(scripting.GetForeignAPI());
     scripting.GenerateEngineBindingsFile();
+
     scripting.SetMainScript(instance, "game/game.wren");
 
     return instance.Run();
