+++ engine/scripting/public/main_script.hpp
@@ -7,31 +7,36 @@
 class MainScript
 {
 public:
-    MainScript() = default;
-    ~MainScript() = default;
-
-    NON_MOVABLE(MainScript);
-    NON_COPYABLE(MainScript);
-
     // Sets the main execution script for Wren
     // This script must define a "Main" class with the following static methods:
     // static Start(engine) -> void
     // static Update(engine, dt) -> void
-    void SetMainScript(wren::VM& vm, const std::string& module, const std::string& className);
+    // static Shutdown(engine) -> void
 
-    // Initializes the main entry point class for Wren
-    void InitMainScript(Engine* e);
+    MainScript(
+        Engine* engine,
+        wren::VM& vm,
+        const std::string& module,
+        const std::string& className);
 
-    // Calls the script main update function
-    void Update(Engine* e, DeltaMS deltatime);
+    ~MainScript();
+
+    NON_MOVABLE(MainScript);
+    NON_COPYABLE(MainScript);
+
+    // Calls the scripts main update function
+    void Update(DeltaMS deltatime);
 
     // Checks if the script can be run without running into an error
     // Either the script is not set or the previous run threw an exception
     bool IsValid() const { return valid; }
 
 private:
+    Engine* engine {};
+
     bool valid = false;
     wren::Variable mainClass {};
     wren::Method mainUpdate {};
     wren::Method mainInit {};
+    wren::Method mainShutdown {};
 };
