+++ game/game.wren
@@ -1,5 +1,5 @@
 import "engine_api.wren" for Engine, TimeModule, ECS, Entity, Vec3, Quat, Math, AnimationControlComponent, TransformComponent, Input, Keycode, SpawnEmitterFlagBits, EmitterPresetID
-import "movement.wren" for PlayerMovement
+import "gameplay/movement.wren" for PlayerMovement
 
 class Main {
 
@@ -72,6 +72,10 @@ class Main {
 
         __rayDistance = 1000.0
         __rayDistanceVector = Vec3.new(__rayDistance, __rayDistance, __rayDistance)
+    }
+
+    static Shutdown(engine) {
+        System.print("Shutdown script!")
     }
 
     static Update(engine, dt) {
