+++ game/tests/wren_main.wren
@@ -2,10 +2,14 @@ import "Engine.wren" for Engine
 
 class ExampleMain {
 
-    static Start() {
+    static Start(engine) {
     }
 
     static Update(engine, dt) {
         System.print("%(dt)")
+    }
+
+    static Shutdown(engine) {
+
     }
 }
